+++ WHATSNEW.md
@@ -1,9 +1,10 @@
 ### Added
 
+- Add timeout option to CLI command fgt backup
+
 ### Changed
 
 - Unpin typer version and fix cli tests
 - Refactor tests
-
-### Removed
+- Change default timeout for fgt backup to 60 seconds